Split Numeric VST into Longs and Doubles
The goal of this change is to let numeric script values sources know if
they are floating point or not.  Previously, this was done by plumbing
the ValueType deep into the ValuesSource.  Now, instead, pass in a
boolean, but that in turn requires the VST to know if it's floating
point or not.  Thus this split.

+++ modules/aggs-matrix-stats/src/main/java/org/elasticsearch/search/aggregations/matrix/ArrayValuesSourceParser.java
@@ -31,8 +31,9 @@ public abstract class ArrayValuesSourceParser<VS extends ValuesSource> implement
 
     public abstract static class NumericValuesSourceParser extends ArrayValuesSourceParser<ValuesSource.Numeric> {
 
+        // NOCOMMIT: Long Support? IDK what we should do here.
         protected NumericValuesSourceParser(boolean formattable) {
-            super(formattable, CoreValuesSourceType.NUMERIC, ValueType.NUMERIC);
+            super(formattable, CoreValuesSourceType.DOUBLE, ValueType.NUMERIC);
         }
     }
